+++ R/tab_create_modify.R
@@ -1288,6 +1288,10 @@ tab_spanner_delim <- function(
           )
 
         # Set the spanner with a call to `tab_spanner()`
+        existing_spanner_ids <- dt_spanners_get_ids(data = data)
+        if (spanner_id %in% existing_spanner_ids) {
+          spanner_id <- random_id(n =5)
+        }
         data <-
           tab_spanner(
             data = data,